[#3705] Add option to completely replace map location page markers
Map location marker style now includes an option for configuring
the map marker icon class, so it can now be completely replaced
in a style if desired.

This also moves the marker rendering code into a separate method
in `MapLocationControlIcon` to make it easier for sub-classing
markers to replace the rendering while retaining the other helpful
parts of the class.