[#3730] Add additional type-specific filters
Adds additional filters defined by specific types. Each document
`TypeDataModel` can define a `compendiumBrowserFilters` property
that returns a Map containing the additional filters that can be
used to filter on that type. There are three types of additional
filters currently defined:

1. `boolean` filters on whether a single value is `true` or not
2. `range` filters a numeric value based on a minimum or maximum
3. `set` filters on a list of values (such as sub-types or
   properties)

For these filters a new `<filter-state>` custom element was added
that allows tracking positive or negative filtering. So if the
filter is on the "magical" property, then a positive state would
only show magical items while the negative state would not show
any magical items.

When multiple actor or item types are selected an intersection of
the provided additional filters is found and only those are shown.
This ensures that only relevant filters are displayed but allows
filters that are shared across types to be handled properly (so
you can filter all physical items based on their rarities, but
only weapons can be filtered on "Weapon Type").