[#3717] Fix appended numbers not being incremented for placed tokens
Since the core's code for appending numbers in `getTokenDocument`
looks at the tokens currently in the scene, and token placement
creates tokens in bulk, this modifies the generated numbers after
the call to `getTokenDocument` to adjust them based on the index
of the placed token.

To that end, this adds a `index` value to `PlacementData` which
has a total index across all placements and an index for each
unique original token, to avoid numbering issues with placement
of mixed groups of actors.