[#3730] Implement initial version of the Compendium Browser
Adds the initial structure of the compendium browser with the
ability to filter by document class (e.g. Actor or Item) and by
type (e.g. NPC, Spell, Weapon). The browser will search all
available compendiums and display all matching documents based
on the types selected.

The compendium browser can be invoked using a button on the
compendium packs sidebar tab, which opens a compendium browser
with no restrictions, or through the Add Race/Background/Class
buttons on the character sheet, which open a compendium browser
instance restricted to those specific types.

When opening a compendium browser programatically, certain filters
can be locked to allow for only viewing a subset of documents.
This locking will be used when invoking the browser from character
creation, from a Choose Items advancement, or when selecting
spells.

The compendium browser has a `fetch` method which is responsible
for getting results from across compendiums. This method filters
compendiums by the appropriate document class and if more specific
types are provided, adds additional compendium filtering using a
new `types` flag that can be defined on packs in the manifest
indicating which types can be found in that compendium. If the
flag isn't set, then that compendium will be searched regardless
of type, but it should reduce the number of compendiums that need
to be searched if these flags are used widely.

The `fetch` method then indexes each compendium using explicitly
provided index fields as well as any fields that are defined in
the filters. Once the index is returned for each compendium it
uses the new `Filter` system to quickly narrow down the results.
If full documents are required, only after this initial filtering
will it fetch those to avoid loading more than necessary.

As a final step it flattens down the results into a single list
and sorts using the provided sorting key path or method, falling
back on alphabetical by name if no sorting is specified.

The browser is implemented using AppV2 which is only available in
Foundry V12, so it is not available when running in V11, but stubs
were used to ensure the rest of the system is still available in
the older version.