[#1964] Implement data structures for storing activities
Introduces the templates and data structures for storing, fetching,
and updating activities on items.

- `ActivityTemplate`: Adds activities and uses to items
- `ActivitiesField`: Mapping field that stores activities and
  initializes as a custom `Collection`.
- `ActivityField`: Field inside the mapping field that stores an
  individual activity, initializing as a data model of the correct
  type.
- `ActivityCollection`: Subclass of `Collection` that organizes
  activities and provides a `getByType` method to retrieve a list
  of activities of a certain type.

In the process this splits `fields.mjs` into separate files because
the original file was getting quite long.

-----------------

This introduces the base activity type, `UtilityActivity`, which
doesn't have much behavior of its own and can be used for when an
item doesn't have a more specific activity it requires. For example
the spell "Mage Armor" would have a `UtilityActivity` that is just
responsible for applying the active effect to the caster.

In order to support the necessary inheritance structure, the base
behavior of all activities is implemented in `ActivityMixin`. This
mixin doesn't have much behavior now, but will eventually handle
activing the activity and consuming uses.

-----------------

A few modifications have been made to `Item5e` in order to support
CRUD operations on activities using the same API that is used for
other documents (e.g. calling
`item.createEmbeddedDocuments("Activity", [{ type: "utility }])`
to create a new activity). This involved adding some additional
behavior to `getCollectionName` and `getEmbeddedCollection` to
recognize special collections defined withing system data.

In addition, `ActivityMixin` inherits from `PseudoDocumentMixin`,
a new set of base methods to implement these CRUD operations.
Eventually `Advancement` should be modifies to use this new setup
so that the `createAdvancement` and related methods on `Item5e`
can be removed.