[#3730] Add selection system for choosing options from browser
Adds a selection interface and API for making choices within the
compendium browser and returning those results to another part of
the application for further use. This uses the new API for the
controls for adding race, background, and classes to a character
sheet.

The Compendium Browser application options get a new `selection`
object with the minimum and maximum number of items that can be
selected. If either of these are defined then checkboxes appear
in the results list and a new footer which displays the number
of selected documents and contains the submit button.

To make this easy to interact with a pair of new factor methods
have been added: `select` and `selectOne`. The `select` method
takes the provided options and returns a set of UUIDs or `null`
if no selection was made. The `selectOne` wraps that method and
sets the minimum and maximum both to `1` and then returns a single
UUID string rather than a set.