[#2623] Separate ability score locking from fixed improvements
Previously if you granted an ability score a fixed improvement it
also locked that score so it cannot be improved further. This
moves the list of locked ability scores into a separate list so
scores without fixed improvements can be locked, and scores with
fixed improvements can still be improved further.

This is a slightly breaking change because fixed improvements that
were previously locked will no longer be so.