[#3799] Add option to register source books directly in manifest
Adds a registration step for modules that detects declarations in
active modules and automatically integrates them during system
preparation.

The only declaration currently supported is `dnd5e.sourceBooks`,
which declares source books in the same format as the `sourceBooks`
config object in the system. The key is a abbreviated name for the
source and the value is the full name, which can be a localization
string.