[#2201, #2550, #2551] Uncap XP tracking & add epic boon calculation
Removes the cap for maximum XP. When the player's level is maximum
it will calculate the number of "boons earned", which means the
number of 30,000 chunks of XP they character has earned above the
maximum XP. This will be displayed on the sheet instead of the
XP threshold for the next level.

The XP required to earn an epic boon can be set using
`CONFIG.DND5E.epicBoonInterval`.