[#1964] Remove extra API code, adjust migrations, remove recharge
Based on feedback this removes the code to make the activities
match core's CRUD methods and instead implements its own API.
Most of `PseudoDocumentMixin` is kept to allow for behavior to be
shared between Activity and Advancement types (though moving
Advancement to that mixin will wait for a future PR to reduce
the size of this PR).

Tweaked the uses migration to fix some bugs and removed the
attempt to migrate uses `value` to `spent`, which will be handled
as part of a concrete migration once those are added. Also added
uses preparation to derive `value` and provide shims for `recharge`
which has been removed.