[#1964] Add basic activity configuration sheet
Adds a new ApplicationV2-based sheet for editing activities. This
sheet is designed to work with the existing `UtilityActivity` and
can be easily subclassed for other activty types.

Note: Due to the ongoing item sheet re-design, no work has been
done to design this sheet to match the system design, only to get
the underlying functionality and UX working.

The sheet consists of three primary tabs, the first of which is
the Identity tab containing the activity name and icon. This tab
will probably be eventually expanded to include a description and
visibility settings once they are added.

The Activation tab contains the bulk of the details with three
sub-tabs. The Time tab contains activation time and condition and
the duration of the activity.

The Consumption tab contains a list of consumption targets,
allowing any number of consumption targets to be configured and
supporting scaling of consumption. The scaling section indicates
whether scaling is allowed (beyond default scaling from spells)
and allows for setting a scaling cap. Uses indicates the spent
uses and a formula for the maximum uses. Use recovery is another
list of recovery profiles, allowing for complex recovery such as
recovering one charge on a short rest and all on a long rest.

Finally the Targeting tab contains details on the range, affected
targets, and area of effect. For area of effect, the dimensions
displayed and their labels will change depending on the selected
area type so you only see the relevant sizes.

The final primary tab, Effect, isn't currently implemented. For
the base activity sheet this tab will contain controls for
selecting what active effects can be applied by this activity. For
other sheets this will display details on the type of saving throw
the must be made, attack to hit modifiers, the damage to apply,
creatures to summon, or enchantments to apply.