[#1964] Add applied effects data to activities & sheet
Adds list of effects for an activity to apply. These effects are
stored as an array of objects with IDs in case the system needs
something like level restrictions in the future (like with
enchantments).

Configuring these applied effects is handled through the Effect
tab of the activity sheet. There is a list of currently applied
effects (unstyled currently) with controls for opening and
deleting effects. There is a button for adding new effects
directly rather than returning to the item sheet, and a
`<multi-select>` allowing for effects that are already on the item
to be selected.