[#1964] Add attribute & material consumption & recharge recovery
Adds full support for configuring attribute & material consumption
including a list of targets when embedded. Also adjusts how item
uses targest are displayed to list charges remaining or for a
period in the same way as they are presented on the current item
sheet.

Adds the recharge recovery period that displays a fixed list of
recharge rolls (e.g. "Recharge 4–6") to select.