[#3910] Add `AttackActivity` data and sheet
Introduces the attack activity which contains data on attack
types, to hit modifiers, and damage parts. This adds the sheet
for the attack activity which includes some extra details on the
Identity tab for the attack type & classification and the to hit
and damage parts on the Effect tab.

Note: Does not contain any functionality for rolling attacks or
damage.