[#1964] Add method to localize schemas in `ArrayField`s
Adds a static `localize` method to activities which performs
pre-localization tasks. During the `i18nInit` hook the system
iterates over all activities and calls this method.

This method calls the normal `localizeDataModel` method on the
activity itself and then calls `Activity#_localizeSchema` to
localize any `SchemaField`s in arrays (this method is necessary
because core's `localizeSchema` method is private).