[#3914] Add `EnchantActivity` data and sheet
Copies over the data structure from `EnchantmentField` and the UI
from `EnchantmentConfig`.

Introduces a new `enchantment` Active Effect type and added a
migration to automatically convert effects with the old flag
to the new type.

The data that was once stored in the flags on the old enchantment
effects has been moved to the `effects` object on the enchant
activity, which seems like a more logical place to store this
information than on each enchantment effect because it has to do
with the application process, not the final enchantment. When
activity migrations are performed the old data in flags will be
moved to the newly created `EnchantActivity`.

**Note:** Does not contain the functionality for enchanting items.