[#1964] Add migrations from item data to activities
Adds migrations to handle creating the default activity on items
based on the old `ActionTemplate` and `ActivatedEffectTemplate`
data. These migrations are performed during the item migration
process so it has access to flags and effects data.

The item calls into `ActivitiesTemplate` which determines whether
the migration should be called at all. It will only perform the
migration if it is a full-data migration, if the item has either
`actionType` or `activation.type` set, and `system.activities` is
empty. This then determines the proper activity type based on
the `actionType` and calls the `migrateInitialActivity` method
on its data model.

`BaseActivityData` handles the base level migration moving all of
the shared activity data over and then passes that object to
`transformTypeData` so specific activity types can handle any
data they need.