[#1964] Remove Action & Activated Effect data & add shims
All of the item types in the system that mix in `ActivityTemplate`
no longer use `ActionTemplate` or `ActivatedEffectTemplate`. To
avoid breakages they now call `_applyActivityShims` which adds
shims to all of the old properties from those templates where
possible.

This PR also includes changes to `WeaponData` that includes
base damage & versatile damage as well as changes to `SpellData`
that includes activation, duration, range, and target objects.

A few additional data changes had to be made to due to gaps:
- `cover` and `crewed` have moved from `ActivatedEffectTemplate`
  to `MountableTemplate`
- `chatFlavor` has been added to all activities and is now used
  in the created chat messages
- Critical bonus formula has been added to attack & damage
  activities, and critical threshold has been added to attack
  activity

Currently the preparation that was performed in
`prepareFinalActivatedEffectData` hasn't been re-implemented, but
most of that will be handled by `BaseActivityData`. This means
that a few labels won't display on legacy item sheets.