[#1964] Migrate weapon base & versatile damage
Migrates the first damage part on a weapon to its base damage
and the versatile formula if present into the versatile damage.
If `@mod` is included it will be automatically stripped because
that will be added automatically for weapon's base damage, and
the type isn't stored in the versatile damage object because it
will also match the type of the base damage.

Modifies the initial advancement creation for `AttackActivity`
to separate out the first damage part from the rest. The first
damage part is kept to be the item's base damage while the rest
are added to the activity.