[#1964] Migrate spell activation & target
Performs minor migration on spell activation to rename `cost` to
`value` and adds a shim for the old value.

Also migrates the old target data to the new format with several
shims to support accessing the old values as best as possible.

Modifies the initial advancement creation to avoid populating
these values on the created activity if the item type is a spell.