[#1964] Prepare final formulas and labels
Implements that final formula resolution and other preparation for
duration, range, and target fields on spells and activities, and
in the uses field everywhere. Some behavior previously implemented
in `ActivatedEffectTemplate` to resolve formulas is now in the new
`prepareFormulaValue` utility function.

This also re-implements the label generation previously handled
in `ActivatedEffectTemplate` for spells and weapons.