[#1964] Copy inferred data from items to activity, add overrides
During activity preparation this copies data from the item over to
the activity unless the `override` option is set. If the original
data couldn't be found on the item, then the override checkbox
won't appear and the activity values will be used.

After the copying occurs a copy of the activity data is saved in
`_inferredSource` so that the activity sheet has a way to show
inferred data before any final formula resolution occurs.

The sheet then selects the item data or activity data based on the
override checkbox, and if override isn't set it will disable the
whole `<fieldset>`.