[#1964] Add inferred attack data, fix bug in save sheet
Infers attack type & classification from the containing item and
displays a label indicating the inferred value if no overriding
value is specific on the `AttackActivity`. Also displays the
potential ability options that might be used for the attack.

When base damage is included, this adds the weapon's base damage
to the list of parts but disables its form elements, so users can
view the weapon's damage on the activity sheet.

Also fixed a bug on the `SaveSheet` causing the incorrect DC
calculation value to be displayed.