[#1964] Remove activation-based preparation from items
This should resolve any data preparation warnings caused by data
that has been moved into activities. This means that most labels
are now generated on an activity's `label` object rather than on
the item, unless it is something like a spell or weapon that has
certain intrinsic data.

The handfull of type-specific label generation that is left on
items has been moved into data models to clean up `Item5e`.

The `ability` data for `AttackActivity` and `SaveActivity` has
been moved into the `attack` and `save` objects respectively to
make room at the top level for an `ability` getter which is used
to set `@mod` in roll data.

A number of the getters on `Item5e` have been marked as deprecated
if they call into deprecated getters on the data model.