[#3910] Move rollAttack to AttackActivity
Adds the Attack button to attack activities and the `rollAttack`
method, which replicates the behavior of that method from `Item5e`
but uses the new rolling API. Internally this constructs config
objects according to the new API to call the `preRollAttackV2`
hook, and then translates that back into the old configuration
format for calling the deprecated hooks and the roll constructor.

A new `criticalThreshold` getter on `AttackActivity` now handles
most of the functionality of the old getter on `ActionTemplate`.
The `criticalThreshold` getter on `ItemDataModel` replaces
`_typeCriticalThreshold` which is no longer required (but is still
checked for backwards compatibility).

A few of the signatures of the new rolling hooks have changed. The
pre hooks no longer pass the activity as the first parameter
because that is included as part of `rollConfig`. The post hooks
now include the activity as part of the second data object, not
the first value. These changes should help in situations like
enrichers where you might be rolling attacks or damage without
an associated activity.