[#3912] Implement damage rolls for Damage & Heal activities
Adds the `rollDamage` method to the activity mixin which provides
a shared method for all activity types that roll damage. This
calls into `getDamageConfig` and `processDamagePart` which provide
places for individual activity types a place to configure the data
provided to damage rolls.

This PR implements rolling damage for `DamageActivity` and
`HealActivity`, and leaves out some parts of the damage data
preparation that are only relevant to attack damage rolls.

Note: Currently just falls back to using the first of the provided
damage types.