[#1964] Implement initial activity usage dialog
Adds the initial activity usage dialog which re-implements the
item usage dialog in ApplicationV2. The new setup should be more
flexible to allow for specific activity types like Enchant and
Summon to add their own controls. It also now re-renders when
changes are made allowing for dynamic inputs and messages.

The dialog currently doesn't implement the consumption warnings
from the original dialog pending the implementation of the
consumption system, nor does it re-scale the item when changes
to scaling are made.