[#1964] Properly scale items when activated
Changes how scaling occurs to be able to handle scaling on all
items, not just spells. It does this by adding the `dnd5e.scaling`
flag which contains the number of steps an item was activated
ahove its base level. So for a spell cast at its base level, this
would be `0` and it would go up `1` for each level it is upcast.

This value is made available in roll data in two ways. For spells,
their roll data is automatically updates to `@item.level` always
coresponds to the proper spell level without having to update
that value when cloning the item. There is also the `@scaling`
value which is the scaling value and `@scaling.increase` which is
the increase.

So for a spell that targets one creature at its base level and one
additional creature for each slot above base level, you would just
enter `@scaling` in the targets field. For a spell that starts at
30 foot range that increase by 10 feet per level you could enter
`30 + (10 * @scaling.increase)` into the distance field. This
should make scaled values easier to enter without having to
subtract the base level from the upcast level like currently.

Cantrips behave a bit differently than normal spells. They
override the `scalingIncrease` value provided by `ItemDataModel`
setting their scaling increase to a fixed value based on character
level. This allows cantrips to scale their damage properly
without having to set a specific cantrip scaling mode.

There is some extra logic to handle the `spellLevel` flag provided
by spell scrolls and tattoos, but the spell scroll creation code
will have to be revisited to ensure it creates the proper
activities.