[#3983] Add weapon masteries to weapons & characters
Adds configuration for weapon masteries and adds them to weapon
sheets. Also adds weapon mastery tracking to character data and
displays them in the weapons list using a symbol similar to
resistance bypasses.

To configure these masteries for a character, this implements a
new version of the `TraitSelector` application built in AppV2.
This application will hopefully be the basis for new versions of
the config applications for actors.

Also adds a `splitSemicolons` utility method to handle splitting
semi-colon-separated lists and removing empty entries.