[#1866, #4011] Update ammo types, add ammo damage & type to weapon
Updates the names for ammunition types and add "Firearm Bullet"
type to bring everything up to date with the 2024 rules.

Adds damage data to consumable items that is only visible for
ammunition. This damage has an additional option defining whether
it supplements the weapon damage or replaces it.

Adds a "Ammunition Type" field to weapons defining what type of
ammo they use. This data will be used in a followup PR to add a
ammo selection UI to the attack roll dialog that only lists ammo
that can be used with the weapon.

Also modified the damage section on weapons to add a hint about
how the weapon damage is used and how `@mod` isn't necessary.
Plus hide damage types for versatile damage since the system
just fetches those from the base damage. Finally it displays a
placeholder for the number and die fields in versatile damage to
indicate that those fields can be inferred automatically.

Closes #1866
Closes #4011