[#1039, #3268, #3314] Add ammo selection, consumption, & bonuses
Adds a list of usable ammunition types to the attack dialog
determined by the ammunition type on the weapon. When one of
these is selected any magic bonus it has is added to the attack
roll and a piece of ammunition is consumed. If the consumption
destroys the ammunition then it is stored in the roll message to
ensure it can be used for damage calculation.

When rolling damage the ammunition used on the last attack is sent
to the `rollDamage` method. If the ammo has its own damage it
either replaces the weapon's base damage or adds a new damage part.
The ammo's magical bonus is added to the base damage.

Closes #1039
Closes #3268
Closes #3314