[#1964] Migrate uses, fix recovery on charges items w/o formula
Performs a concrete migrate for `uses.value` because it relies on
the resolved max uses formula. The involved some changes to how
migrations are called to pass the fully prepared item through the
process.

Also fixed a migration issue for items that have charges but not
a recovery formula set getting a recovery period.