[#4024] Add `CheckActivity` type for ability & tool checks
Adds the check activity for rolling ability, skill, or tool checks.
It must specify an ability and can optionally specify a skill or
tool that can contribute proficiency, if the player has it.

If added to a tool item, then it can infer the ability and the
associated proficiency from the tool, along with any special bonus
the tool provides.

Closes #4024