[#1964] Move concentration into activity duration
Adds `concentration` into the duration data on activities and
removed it as a property on all items except spells. This allows
concentration to be handled on a per-activity basis rather than
for an item as a whole.

For spells, they get the derived `duration.concentration` value
based on their properties and that will be inferred on the activity
unless the duration is overridden.