[#1964] Adjust styling of usage, switch to <dialog> element
Adjusts the styling of the activity usage dialog to match the
V2 system styling more closely. This includes a redesigned
range input which wasn't previously styled.

Adds a subtitle to the header of all V2 sheets configurable
via `options.subtitle` or the `subtitle` getter. This will be
hidden when the application is minimized.

Also changes the outer element of the usage application to be a
`<dialog>` element and the content tag to be `<form>`.

Note: Resource consumption currently isn't styled awaiting a
followup PR that makes more significant changes to those options.