Add `autoPosition` option to roll dialogs
`BasicRollDialogConfiguration` now accepts the `autoPosition`
option. When used, this positions the dialog relative to the
button used to activate it. This replaces the manual positioning
performed in `rollAttack` and `rollDamage` to place the dialog
near the chat.

Places the dialog to the right and slightly below the button
clicked unless there isn't enough space, in which case it shifts
to the left side.