[#4043] Refresh Create Scroll UI, allow custom spell values
Reworks the Create Scroll using ApplicationV2 in a way that allows
for dynamic updates of the form. It now displays fields for spell
save DC and attack bonus with placeholders updating dynamically
when the spell level is changed.

These fields can also be set manually, overriding the default DC
of spells. If using the modern rules version and create a scroll
directly from a spell in an actor's inventory, then these values
will be initial set based on that actor's spellcasting stats.

Closes #4043