[#1012, #1229] Display classes providing spells on spell sheet
Modifies the `SpellListRegistry` to keep track of what spell lists
provide a spell in addition to spell lists by type. This allows
spells to retrieve a list of spell lists they appear on and turn
it into a list of classes to be displayed on the spell sheet.

Part of this work involves fetching the name of the class from the
item registry rather than using the name of the spell list itself.
In order to get around the loading time involved a new API has
been added to the registry that allows defering a callback until
all registries have fully loaded.

Closes #1229