[#3836, #3853] Add Has Spellcasting & Class filters to subclasses
Adds a "Has Spellcasting" filter to subclasses in the compendium
browser indicating the subclass provides spellcasting. Also adds
the ability to filter by the class to which the subclass belongs.

To support the second filter, this introduces the `ItemRegistry`
to the system registry that scans compendiums and loads all
classes registered in different compendium packs. These are
grouped by identifier and different versions of a class in
different packs are consolidated with their UUIDs listed in the
`sources` list. This only supports classes at the moment but
could easily be expanded to incude subclasses, backgrounds, or
species, which maybe useful to work on spell lists.

Closes #3836
Closes #3853