[#1964] Re-implement effect application for activities
Re-implements the effect application UI for activities using a
custom element like damage application. This allows for applying
effects to targeted or selected tokens. The top part displays the
targets (with checkboxes to control whether they are applied when
using the "Targeted" mode) and below that is a list of effects to
apply.

Refactor some shared code from `DamageApplicationElement` into
`TargetedApplicationMixin` to handle targeting tokens.

Also makes a change to the damage application UI to display token
names rather than actor names.