[#1407] Add Subclass advancement & selection interface on sheet
Introduces a new subclass advancement! This advancement doesn't do
anything on its own other than indicate the level at which the
class takes a subclass due to limitations in the advancement
system.

When a player reaches the level of the subclass advancement, their
features tab adds a "Add Subclass" button where the subclass icon
will eventually appear. Clicking this opens the compendium browser
with the subclass list locked to the current class's identifier.

While this doesn't fully implement what was originally desired
from the subclass advancement, this workflow should work pretty
well to fill in a major gap in the system's level-up workflow.

Closes #1017
Closes #1407