[#3983] Add mastery selection & display mastery property in chat
Adds the `traits.weaponProf.bonsuMasteries` data to characters
that contains extra masteries that can be applied to any weapon
in which they have mastery, in place of the weapon's existing
mastery.

When making an attack roll using a weapon with which the character
has mastery, a dropdown will appear in the dialog if there is more
then one mastery option, otherwise the first mastery will be
selected automatically for the attack.

The mastery used will appear in the attack roll chat card and if
a reference was provided for that mastery, it will be a link so
players can hover and see a description of the mastery property.

Closes #3983