[#4079] Handle limited uses on granted spells
Handles setting the limited uses on spells granted through an
advancement in a way that respects activities. To support all the
desired workflows, a "Requires Slot" checkbox has been added to
the spell config which controls how the limited uses are applied.

- If preparation mode doesn't require a spell slot by default or
  "Requires Slot" is checked, then consumption is added to all
  activities on the spell marked to "Consume Spell Slot"
- If preparation mode requies a slot and "Requires Slot" isn't
  checked, then each activity marked to "Consume Spell Slot" is
  duplicated, given consumption, removed spell slot consumption,
  and has "Free Casting" added to its name

This combination should handle all of the standard configurations
needed for official content.

This PR also reintroduces `system.ability` to spells as a means
of overriding the default ability from the class or actor. This
field only appears for embedded spells and shows the default
ability when not set.

Closes #4079