[#2508] Add spell list filter to compendium browser
Adds a new filter for spells to the Compendium Browser that allows
filtering by a specific spell list. This involves some tweaks to
how spell lists are stored in the registry to allow for creating
a list of all lists easily.

This new filter allows for easily finding all the spells for a
specific class across all modules that register spell lists. It
also allows doing things like finding spells that are on the
spell list of one class but not on any others.

Closes #2508