[#4024] Rework Check Activity to support multiple checks
Changes the `check.associated` value on `CheckActivity` to be a
set of associated values rather than a single value. If multiple
associated values are set then more than one button will appear in
the chat card.

Setting an explicit ability for the activity will cause all of the
created checks to use that ability, otherwise they will attempt to
determine the best ability to use (skills from the default value
in config, tools from the tool item or character data).