[#3803] Add rider activities to enchantments
Allows activities to be applied to the item alongside enchantments
like effects and items are currently.

The previous method of using a flag on effects to mark them as
riders wasn't practical in this case due to activites not having
flags, so this stores the rider activities and effects in a new
flag on the item itself. A migration is included to move the
`rider` flag from the effects to the item flags.

Closes #3803