[#2138] Refactor `rollRecharge` to use `BasicRoll` & new hooks
The roll recharge method now makes use of the new configuration
structure and rolling hooks to be consistant with other rolls in
the system. It also gains a `dnd5e.postRollRecharge` hook that
fires after the usage updates have been performed. A similar post
hook has been added to `rollHitDice` so that all rolls that
perform updates now have a `post*` hook after those updates are
complete.

In order to support modifying the chat message flavor based on
the result of the recharge roll, a `preCreate` callback has been
added to `BasicRollMessageConfiguration` that is called right
before the message is creating, allowing for any last-minute
changes to the message. This hook will also be very helpful when
attack rolls are hit by the roll refactor.

A small tweak has been made to the config and hooks to consistantly
use `subject` for the Document performing the roll, rather than
using `origin` in the initial config and a document-type specific
key in the other hooks.