[#1539] Add identifier field to all items
Adds `identifier` field to `ItemDescriptionTemplate` so it is
added to all items. For classes, subclasses, backgrounds, and
species it can still be edited from the details tab, but for
other item types it is now available in the Configure Source app.

Also makes a small change to how identifiers are created to handle
slashes in the middle of words. This avoids an issue where a name
like "Blindness/Deafness" would result in an identifier of only
one word (e.g. `blindnessdeafness`) and instead adds a dash in
place of the slash (e.g. `blindness-deafness`).