[#2138] Refactor `DamageRoll` to use new rolling API
Reworks parts of `DamageRoll` to match the design of `BasicRoll`
and removed a number of methods not needed anymore. Deprecated the
`damageRoll` standalone method and changed it internally to use
the new `DamageRoll#build` method. Finally adjusted `rollDamage`
to use the new build process.

Implemented a new roll configuration dialog for damage rolls that
allows for selecting damage type and improved the base dialog to
support entering situational bonuses for each damage part, rather
than for just the first one.