[#1539] Use migration to set identifier on items
Sets the identifier for items using a one-time concrete migration.
Avoided an in-memory migration because an empty identifier is
considered valid.

Adds a `bypassVersionCheck` flag to migrations that allows
skipping the check that prevents certain migrations from being
applied to items updated in `4.0.0` or later. This is a flag to
allow for re-migrating items that have already been migrated to
`4.0`.

Fixes minor issue where `_migrateItemUses` causes item updates
every time even if flag isn't present.