[#4195] Add modern style option for class/subclass summary pages
Adds the ability for class and subclass summary journal entry
pages to be formatted in a way that matches the 2024 PHB, rather
than the 2014 PHB. This incudes moving the traits into a table
above the introduction and prefixing feature names with their
level.

The styling is decided automatically based on the rules version
specified on the class/subclass item, but can be overridden for
each journal page if desired.

Closes #4195