[#4135] Pre-create `MappingField` data before applying effects
Fix bugs caused by applying an active effect targeting a value in
and entry in `system.tools` when the actor doesn't currently have
that entry. The bug was caused by the whole object not being
created and other parts of the code expecting it to exist.

This solves the problem by checking if an active effect is trying
to apply to an `MappingField` entry that doesn't exist yet, and if
so it will first create the initial entry before continuing with
the application process.

Closes #4135