[#4276] Fix effect bug caused by `persistSourceMigration`
When the `persistSourceMigration` flag was enabled it updates an
item with `diff: false`, but when an effect migration is applied
the update data contains only partial updates for effects. This
would result in the system attempting to update effects without
most of their data leading to an error.

This fix resolves this isssue by manually merging changes with
the original effect data before performing the full merge.

Closes #4276