[#4278, #4279, #4297] Fix concentration from activities
Changes a number of concentration implementation details to fully
work with activities and fix issues with concentration on
non-spell items or activities that offer concentration even if the
base item does not.

Modified `beginConcentrating` and `createConcentrationEffectData`
to accept an activity rather than an item, with a fallback that
selects the first activity and shows a deprecation warning.

Deprecated `ActiveEffect5e#getEffectDurationFromItem` and replaced
it with a `getEffectData` method on `DurationField` which serves
the same purpose.

Modified concentration checks in the activity usage process to use
`requiresConcentration` on the activity rather than the item so
that activites that require concentration on items that do not
still properly trigger.

Also tweaked the flags added to the concentration effect to include
activity and scaling data.

Closes #4278
Closes #4279
Closes #4297